requirements.txt
Flask>=2.0.2
Werkzeug>=2.0.2
tensorflow==2.13.0
keras==2.13.1
h5py==3.4.0
Pillow>=8.0.0
